+++ libs/vedana-etl/src/vedana_etl/pipeline.py
@@ -5,15 +5,21 @@ from datapipe.step.batch_transform import BatchTransform
 import vedana_etl.steps as steps
 from vedana_etl.catalog import (
     dm_anchors,
-    dm_attributes,
+    dm_link_attributes,
+    dm_anchor_attributes,
     dm_links,
+    dm_queries,
+    dm_prompts,
+    dm_conversation_lifecycle,
     edges,
     grist_edges,
     grist_nodes,
     memgraph_edges,
-    memgraph_indexes,
     memgraph_nodes,
-    memgraph_vector_indexes,
+    memgraph_anchor_indexes,
+    memgraph_link_indexes,
+    memgraph_anchor_vector_indexes,
+    memgraph_link_vector_indexes,
     nodes,
     eval_gds,
 )
@@ -21,7 +27,15 @@ from vedana_etl.catalog import (
 data_model_steps = [
     BatchGenerate(
         func=steps.get_data_model,  # Generator with main graph data
-        outputs=[dm_anchors, dm_attributes, dm_links],
+        outputs=[
+            dm_anchors,
+            dm_anchor_attributes,
+            dm_link_attributes,
+            dm_links,
+            dm_queries,
+            dm_prompts,
+            dm_conversation_lifecycle,
+        ],
         labels=[("flow", "regular"), ("flow", "on-demand"), ("stage", "extract"), ("stage", "data-model")],
     ),
 ]
@@ -58,9 +72,19 @@ default_custom_steps = [
 
 memgraph_steps = [
     BatchTransform(
-        func=steps.ensure_memgraph_indexes,
-        inputs=[dm_attributes],
-        outputs=[memgraph_indexes, memgraph_vector_indexes],
+        func=steps.ensure_memgraph_node_indexes,
+        inputs=[dm_anchor_attributes],
+        outputs=[
+            memgraph_anchor_indexes,
+            memgraph_anchor_vector_indexes,
+        ],
+        labels=[("flow", "regular"), ("flow", "on-demand"), ("stage", "load")],
+        transform_keys=["attribute_name"],
+    ),
+    BatchTransform(
+        func=steps.ensure_memgraph_edge_indexes,
+        inputs=[dm_link_attributes],
+        outputs=[memgraph_link_indexes, memgraph_link_vector_indexes],
         labels=[("flow", "regular"), ("flow", "on-demand"), ("stage", "load")],
         transform_keys=["attribute_name"],
     ),
@@ -70,7 +94,7 @@ memgraph_steps = [
     # generate_embeddings is a last processing step, making DataFrame ready for upload
     BatchTransform(
         func=steps.generate_embeddings,
-        inputs=[nodes, memgraph_vector_indexes],
+        inputs=[nodes, memgraph_anchor_vector_indexes],
         outputs=[memgraph_nodes],
         labels=[("flow", "regular"), ("flow", "on-demand"), ("stage", "load")],
         transform_keys=["node_id", "node_type"],
@@ -78,7 +102,7 @@ memgraph_steps = [
     ),
     BatchTransform(
         func=steps.generate_embeddings,
-        inputs=[edges, memgraph_vector_indexes],
+        inputs=[edges, memgraph_link_vector_indexes],
         outputs=[memgraph_edges],
         labels=[("flow", "regular"), ("flow", "on-demand"), ("stage", "load")],
         transform_keys=["from_node_id", "to_node_id", "edge_label"],
@@ -95,7 +119,11 @@ eval_steps = [
 ]
 
 
-def get_pipeline(custom_steps: list):
+def get_data_model_pipeline() -> Pipeline:
+    return Pipeline(data_model_steps)
+
+
+def get_pipeline(custom_steps: list) -> Pipeline:
     pipeline = Pipeline(
         [
             *data_model_steps,